add user to org functionality
- removes some try catches in order to pass the CF response to UI
- moves content-type to calling function which overrides header with token

fixes and reorganizes tests after addData alterations
