+++ __tests__/api/cloudfoundry/cloudfoundry.test.js
@@ -1,18 +1,16 @@
 import nock from 'nock';
-import { cookies } from 'next/headers';
 import { describe, it, expect, beforeEach, afterEach } from '@jest/globals';
 import {
-  getToken,
   getCFOrg,
   getCFOrgUsers,
   getCFOrgs,
-} from '../../api/cloudfoundry';
+} from '../../../api/cloudfoundry/cloudfoundry';
 import {
   mockOrg,
   mockOrgNotFound,
   mockOrgUsers,
   mockOrgs,
-} from './mocks/organizations';
+} from '../mocks/organizations';
 
 /* global jest */
 /* eslint no-undef: "off" */
@@ -79,45 +77,6 @@ describe('cloudfoundry tests', () => {
       nock(process.env.CF_API_URL).get('/organizations').reply(200, mockOrgs);
       const res = await getCFOrgs();
       expect(res).toEqual(mockOrgs.resources);
-    });
-  });
-
-  describe('getToken', () => {
-    describe('when token environment variable is set', () => {
-      beforeEach(() => {
-        process.env.CF_API_TOKEN = 'manual-token';
-      });
-      afterEach(() => {
-        delete process.env.CF_API_TOKEN;
-      });
-      it('returns a manual token', () => {
-        expect(getToken()).toBe('manual-token');
-      });
-    });
-
-    describe('when token environment variable is not set', () => {
-      describe('when auth cookie is set', () => {
-        beforeEach(() => {
-          cookies.mockImplementation(() => ({
-            get: () => ({ value: '{"accessToken":"cookie-token"}' }),
-          }));
-        });
-        it('returns a token from a cookie', () => {
-          expect(getToken()).toBe('cookie-token');
-        });
-      });
-      describe('when auth cookie is not set', () => {
-        beforeEach(() => {
-          cookies.mockImplementation(() => ({
-            get: () => null,
-          }));
-        });
-        it('throws an error when no cookie is set', () => {
-          expect(() => getToken()).toThrow(
-            'accessToken not found, please confirm you are logged in'
-          );
-        });
-      });
     });
   });
 });
