+++ api/api.js
@@ -7,13 +7,15 @@ export async function addData(url, body, options = {}) {
     const reqOptions = {
       method: 'POST',
       body: JSON.stringify(body),
-      'Content-Type': 'application/json',
       ...options,
     };
     const res = await fetch(url, reqOptions);
-    if (res.ok) {
-      return await res.json();
+    if (res.ok || res.status == 422) {
+      const json = await res.json();
+      return json;
     } else {
+      // note: if debugging cloudcoundry responses, it may be helpful
+      // to check the res.json() contents here
       throw new Error(`an error occurred with response code ${res.status}`);
     }
   } catch (error) {
